LayoutTests/platform/ios-sim/fast/forms/input-spinbutton-capturing-expected.txt

Test for a capturing bug of spin buttons in a type=numnber input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL anotherInput.value should be 2. Was 1.
PASS successfullyParsed is true

TEST COMPLETE
